[Windows] Start refactor of SerialKeyEvent* classes
