+++ windows/src/engine/keyman32/Keyman32.vcxproj
@@ -251,7 +251,6 @@
     <ClCompile Include="k32_tsf.cpp" />
     <ClCompile Include="keybd_shift.cpp" />
     <ClCompile Include="keyboardoptions.cpp" />
-    <ClCompile Include="keyeventsenderthread.cpp" />
     <ClCompile Include="KEYMAN32.CPP">
       <AdditionalIncludeDirectories Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -313,6 +312,8 @@
     <ClCompile Include="preservedkeymap.cpp" />
     <ClCompile Include="security.cpp" />
     <ClCompile Include="selectkeyboard.cpp" />
+    <ClCompile Include="serialkeyeventclient.cpp" />
+    <ClCompile Include="serialkeyeventserver.cpp" />
     <ClCompile Include="syskbd.cpp">
       <AdditionalIncludeDirectories Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -373,12 +374,13 @@
     <ClInclude Include="..\..\global\inc\keyman64.h" />
     <ClInclude Include="..\..\global\inc\keymanapi.h" />
     <ClInclude Include="..\..\global\inc\keymancontrol.h" />
-    <ClInclude Include="keyeventsenderthread.h" />
     <ClInclude Include="keystate.h" />
     <ClInclude Include="..\..\global\inc\Km95api.h" />
     <ClInclude Include="..\..\global\inc\rc4.h" />
     <ClInclude Include="..\..\global\inc\registry.h" />
     <ClInclude Include="security.h" />
+    <ClInclude Include="serialkeyeventclient.h" />
+    <ClInclude Include="serialkeyeventcommon.h" />
     <ClInclude Include="syskbd.h" />
     <ClInclude Include="..\..\global\inc\unicode.h" />
     <ClInclude Include="vkscancodes.h" />
