+++ windows/src/engine/keyman32/Keyman32.vcxproj.filters
@@ -120,13 +120,16 @@
     <ClCompile Include="k32_lowlevelkeyboardhook.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="keyeventsenderthread.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="security.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="pch.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="serialkeyeventclient.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="serialkeyeventserver.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -213,13 +216,16 @@
     <ClInclude Include="k32_tsf.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="keyeventsenderthread.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="security.h">
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="pch.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="serialkeyeventclient.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="serialkeyeventcommon.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
