+++ windows/src/engine/keyman32/serialkeyeventclient.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#ifdef USE_KEYEVENTSENDERTHREAD
+
+#include "serialkeyeventcommon.h"
+
+BOOL SignalKeyEventSenderThread(PINPUT pInputs, DWORD nInputs);
+
+#endif
