+++ windows/src/engine/keyman32/serialkeyeventcommon.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#ifdef USE_KEYEVENTSENDERTHREAD
+
+#define MAX_KEYEVENT_INPUTS 256
+
+#define KEYEVENT_WINDOW_CLASS "Keyman_KeyEventConsumerWnd"
+
+#define GLOBAL_FILE_MAPPING_NAME "KeymanEngine_KeyEvent_FileMapping"
+#define GLOBAL_KEY_EVENT_NAME "KeymanEngine_KeyEvent"
+#define GLOBAL_KEY_MUTEX_NAME "KeymanEngine_KeyMutex"
+
+/**
+  The INPUT structure and the KEYBDINPUT structure both vary in size between x86 and x64
+  because of the presence of the ULONG_PTR member dwExtraInfo. Thus we need to maintain an
+  equal sized structure between the two platforms for shared memory, and copy into INPUT
+  structures before sending the input.
+*/
+struct CSDINPUT {
+  WORD      wVk;
+  WORD      wScan;
+  DWORD     dwFlags;
+  DWORD     time;
+  ULONGLONG extraInfo;
+};
+
+struct ConsumerSharedData {
+  DWORD nInputs;
+  CSDINPUT inputs[MAX_KEYEVENT_INPUTS];
+};
+
+#endif
